requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.3
Django==4.2
djangorestframework==3.16.1
pillow==11.3.0
python-decouple==3.8
reportlab==4.4.4
sqlparse==0.5.3
